Moving credits to a new Component
