+++ TikTok/src/components/Credits/index.js
@@ -1,0 +1,1 @@
+export * from "./Credits"
